Add language to taglink CRUD, fix file deletion in translations.

+++ trust_path/libraries/tuskfish/class/Tfish/Content/Model/ContentEdit.php
@@ -133,11 +133,7 @@ class ContentEdit
             return false;
         }
 
-        // Insert the taglinks, which requires knowledge of the ID. 
-        /*if (!$this->saveTaglinks($content['id'], $content['language'], $content['type'], 'content', $tags)) {
-            return false;
-        }*/
-        if (!$this->saveTaglinks($content['id'], $content['type'], 'content', $tags)) {
+        if (!$this->saveTaglinks($content['id'], $content['language'], $content['type'], 'content', $tags)) {
             return false;
         }
 
@@ -167,25 +163,35 @@ class ContentEdit
         $content['media'] = $savedContent['media'];
 
         // Check if there are any redundant image/media files that should be deleted.
-        if (!empty($savedContent['image']) && $savedContent['image'] !== $content['image']) {
-            $this->fileHandler->deleteFile('image/' . $savedContent['image']);
+        if (!empty($savedContent['image']) 
+            && $savedContent['image'] !== $content['image']
+            && $this->soleReferenceToFile('image', $savedContent['image'])) {
+                $this->fileHandler->deleteFile('image/' . $savedContent['image']);
         }
 
-        if (!empty($savedContent['media']) && $savedContent['media'] !== $content['media']) {
-            $this->fileHandler->deleteFile('media/' . $savedContent['media']);
+        if (!empty($savedContent['media'])
+            && $savedContent['media'] !== $content['media']
+            && $this->soleReferenceToFile('media', $savedContent['media'])) {
+                $this->fileHandler->deleteFile('media/' . $savedContent['media']);
         }
 
         // Check if delete flag was set.
         if ($_POST['deleteImage'] === '1' && !empty($savedContent['image'])) {
             $content['image'] = '';
-            $this->fileHandler->deleteFile('image/' . $savedContent['image']);
+
+            if ($this->soleReferenceToFile('image', $savedContent['image'])) {
+                $this->fileHandler->deleteFile('image/' . $savedContent['image']);
+            }
         }
 
         if ($_POST['deleteMedia'] === '1' && !empty($savedContent['media'])) {
             $content['media'] = '';
             $content['format'] = '';
             $content['fileSize'] = 0;
-            $this->fileHandler->deleteFile('media/' . $savedContent['media']);
+
+            if ($this->soleReferenceToFile('media', $savedContent['media'])) {
+                $this->fileHandler->deleteFile('media/' . $savedContent['media']);
+            }
         }
 
         // Upload any new image/media files and update file names. 
@@ -193,7 +199,7 @@ class ContentEdit
         $this->uploadMedia($content);
 
         // Update taglinks.
-        $this->updateTaglinks($id, $content['type'], 'content', $tags);
+        $this->updateTaglinks($id, $content['language'], $content['type'], 'content', $tags);
 
         // Flush cache.
         $this->cache->flush();
@@ -289,6 +295,38 @@ class ContentEdit
             ->fetch(\PDO::FETCH_ASSOC);
 
         return !empty($row) ? $row : [];
+    }
+
+    /**
+     * Checks to see if an image or media file attachment is in use by another translation of this
+     * content object, in order to avoid deleting files that are in service.
+     *
+     * @param string $field Name of file field, ie. 'image' or 'media'.
+     * @param string $filename Name of file.
+     * @return boolean True if there is only one reference to a file, false if there are more.
+     */
+    private function soleReferenceToFile(string $field, string $filename): bool {
+        
+        if ($field !== 'image' && $field !== 'media') {
+            return false;
+        }
+
+        $criteria = $this->criteriaFactory->criteria();
+        $criteria->add($this->criteriaFactory->item($field, $filename));
+        $count = $this->runCount($criteria);
+        
+        return $count === 1 ? true : false;
+    }
+
+    /**
+     * Run the count query.
+     * 
+     * @param   \Tfish\Criteria $criteria Filter criteria.
+     * @return  int Count.
+     */
+    private function runCount(\Tfish\Criteria $criteria): int
+    {
+        return $this->database->selectCount('content', $criteria);
     }
 
     /**
